extract `grug_crypto::identity_digest` to a new crate
