+++ grug/crypto/Cargo.toml
@@ -14,6 +14,7 @@ blake2        = { workspace = true }
 blake3        = { workspace = true }
 digest        = { workspace = true }
 ed25519-dalek = { workspace = true, features = ["batch", "digest"] }
+identity      = { workspace = true }
 k256          = { workspace = true }
 p256          = { workspace = true }
 sha2          = { workspace = true }
