+++ grug/crypto/benches/benchmarks.rs
@@ -5,10 +5,11 @@ use {
     },
     ed25519_dalek::Signer,
     grug_crypto::{
-        Identity256, Identity512, blake2b_512, blake2s_256, blake3, ed25519_batch_verify,
-        ed25519_verify, keccak256, secp256k1_pubkey_recover, secp256k1_verify, secp256r1_verify,
-        sha2_256, sha2_512, sha3_256, sha3_512,
+        blake2b_512, blake2s_256, blake3, ed25519_batch_verify, ed25519_verify, keccak256,
+        secp256k1_pubkey_recover, secp256k1_verify, secp256r1_verify, sha2_256, sha2_512, sha3_256,
+        sha3_512,
     },
+    identity::{Identity256, Identity512},
     p256::ecdsa::signature::DigestSigner,
     rand::{RngCore, rngs::OsRng},
     std::{hint::black_box, time::Duration},
