+++ grug/crypto/src/lib.rs
@@ -1,8 +1,8 @@
 mod ed25519;
 mod error;
 mod hashers;
-mod identity_digest;
 mod secp256k1;
 mod secp256r1;
+mod utils;
 
-pub use crate::{ed25519::*, error::*, hashers::*, identity_digest::*, secp256k1::*, secp256r1::*};
+pub use crate::{ed25519::*, error::*, hashers::*, secp256k1::*, secp256r1::*};
