+++ grug/crypto/src/secp256k1.rs
@@ -1,5 +1,6 @@
 use {
-    crate::{CryptoError, CryptoResult, Identity256, to_sized},
+    crate::{CryptoError, CryptoResult, utils::to_sized},
+    identity::Identity256,
     k256::ecdsa::{RecoveryId, Signature, VerifyingKey, signature::DigestVerifier},
 };
