+++ hyperlane/isms/multisig/Cargo.toml
@@ -20,8 +20,8 @@ grug            = { workspace = true }
 hyperlane-types = { workspace = true }
 
 [dev-dependencies]
-grug-crypto = { workspace = true }
 hex-literal = { workspace = true }
+identity    = { workspace = true }
 k256        = { workspace = true }
 rand        = { workspace = true }
 test-case   = { workspace = true }
